Start weave proxy independently
Updated weave ssh driver so that it starts weave proxy as an
independent command. This allows us to pass commands to it, such as the
TLS settings.

This was required as we needed to pass the proxy the -H command to
instruct the proxy to expose dockers remote API externally