Auto-checkpointing images post-install: configurable
Now the default is to *not* auto-checkpoint the image after install has
completed (which used to be done for entities that were not already 
pure-docker images).

The auto-checkpoint was dangerous for sshd containers, because it 
created the image after the ssh credentials had been put into 
the container. When re-using the image, you’d have to have exactly
the same credentials. But it’s very brittle to make assumptions about
how that was created.