set container links with a map of alias string to target entity  - avoids having to set the `containerName` on the target which means you can deploy multiple instances of linked container groups  - allows a container to be linked to multiple containers
This is currently a breaking change for anyone using the old list of entities links config.

So the following old yaml (showing only the links related config here):
```
services:
- type: docker:mariadb:5.5
  id: db
  containerName: mysql
- type: docker:wordpress:4.5
  links:
  - $brooklyn:component("db")
```

Is replaced with the following, note the containerName in the db container (which was being used as the alias), has been replaced with an explicitly defined alias in the links map.
```
services:
- type: docker:mariadb:5.5
  id: db
- type: docker:wordpress:4.5
  links:
    mysql: $brooklyn:component("db")
```